refactor(provider): extract the timeout-bounded fetch into src/utils/fetch.ts
`withTimeout` is a fetch primitive, not a provider concern, and it was the only
reason `src/utils/provider.ts` exported a helper purely for its test. The
mechanism is unchanged and the seam is what moved: `src/utils/fetch.ts` imports
nothing from `src/utils.ts` or `src/utils/provider.ts`, so the seven
circular-dependency warnings the build already prints are unchanged.

The primitive stays generic. The `LibraryError` with `code: 'TIMEOUT'` is a
constraint of `RpcChannel#errorHandler`, which re-throws anything else as a bare
`Error(message)` and drops own properties, so it now lives at the
`getStarknetProvider` call site and reaches the primitive as a `timeoutError`
factory. The factory is what the body-phase mapping needs too: the deadline
outlives the headers, so `.json()` has to be able to throw the caller's error
type as well, which a `catch` at the call site could not do without duplicating
the response wrapper.

Everything else moves untouched: the deadline clears when `.json()` settles
rather than when the headers arrive, the caller's `signal` is composed with its
own reason and unsubscribed on settle, the timer unhooks that listener itself
and is `unref`'d so a response whose body is never read neither leaks a
listener nor holds the event loop open, and a non-positive timeout returns the
fetch untouched. Non-finite normalization stays in `normalizeOptions`, where it
can fall back to a default instead of disabling the bound.

`utils.fetch` keeps its own AbortController dance. Sharing this one would extend
its deadline over the body read, and the deadline is cleared by the body read:
`parseScoreAPIResponse` (getScores, getVp, validate) and `subgraphRequest` use
`res.text()`, and an external caller of a documented drop-in `fetch` may read
nothing at all. Measured on a local server with a 3s timeout: `.json()` exits at
+17ms, `.text()` at +3004ms, no read at +3002ms, against +15ms for `utils.fetch`
as it stands.
